Maybe fix an obscure Abyss exit bug (CarefulOdds)
meaning that the game considered it to be temporary terrain. A water
nymph was killed in the same action that the exit was spawned. I was
unable to repro the issue, but I believe that the game still considered
the exit to be part of the nymph's temporary terrain. This tries to
clean up that situation.

Closes #2618.